Revert "RTED passes on 64bit"
This reverts commit d5821a31df48ab97397bd245da7057b97be9a394.

+++ projects/RTED/CppRuntimeSystem/MemoryManager.cpp
@@ -87,7 +87,7 @@ bool MemoryType::operator< (const MemoryType & other) const
 bool MemoryType::isInitialized(int offsetFrom, int offsetTo) const
 {
     assert(offsetFrom<offsetTo);
-    assert(((unsigned int)offsetFrom) < size && ((unsigned int)offsetTo) <= size);
+    assert(offsetFrom < size && offsetTo <= size);
 
     for(int i= offsetFrom; i<offsetTo; i++)
         if(! initialized[i])
@@ -99,7 +99,7 @@ bool MemoryType::isInitialized(int offsetFrom, int offsetTo) const
 void MemoryType::initialize(int offsetFrom, int offsetTo)
 {
     assert(offsetFrom<offsetTo);
-    assert(((unsigned int)offsetFrom) < size && ((unsigned int)offsetTo) <= size);
+    assert(offsetFrom < size && offsetTo <= size);
 
     for(int i=offsetFrom; i<offsetTo; i++)
         initialized[i]=true;
